Reduced training steps of TCN, and eliminated float warning

+++ nixtla/models/tcn/tcn.py
@@ -241,7 +241,7 @@ class TCN(object):
                                                                                             self.loss,
                                                                                             training_loss.cpu().data.numpy())
                     self.trajectories['iteration'].append(iteration)
-                    self.trajectories['train_loss'].append(np.float(training_loss.cpu().data.numpy()))
+                    self.trajectories['train_loss'].append(training_loss.cpu().data.numpy())
 
                     if val_ts_loader is not None:
                         loss = self.evaluate_performance(ts_loader=val_ts_loader,
@@ -274,7 +274,7 @@ class TCN(object):
         #End of fitting
         if n_iterations >0:
             # This is batch loss!
-            self.final_insample_loss = np.float(training_loss.cpu().data.numpy()) if not break_flag else best_insample_loss
+            self.final_insample_loss = training_loss.cpu().data.numpy() if not break_flag else best_insample_loss
             string = 'Step: {}, Time: {:03.3f}, Insample {}: {:.5f}'.format(iteration,
                                                                             time.time()-start,
                                                                             self.loss,
